RobotOps_Streamlit_UI/requirements.txt
# RobotOps_Streamlit_UI – run from this directory: streamlit run app.py
streamlit>=1.28.0
numpy>=1.24.0,<3
pandas>=2.0.0,<2.3
plotly>=5.18.0
pydantic>=2.0.0
paho-mqtt>=2.0.0
# Optional PDF export (uncomment if needed)
# fpdf2>=2.7.0
# reportlab>=4.0.0
